+++ R/mod_raw_table.R
@@ -3,15 +3,17 @@ ui_raw_table <- function(id, label, ...) {
   tabPanel(
     label,
     br(),
-    fluidRow(
-      column(3, ui_varselect(ns("vars"), "Variables")),
-      column(3, shinyWidgets::pickerInput(
-        ns("nrow"), "Rows per page", list("All", "100", "50", "10")
-      )),
-      column(3, downloadButton(ns("download"), "Download"))
-    ),
-    ...,
-    DT::dataTableOutput(ns("data"))
+    fluidPage(
+      fluidRow(
+        column(3, ui_varselect(ns("vars"), "Variables")),
+        column(3, shinyWidgets::pickerInput(
+          ns("nrow"), "Rows per page", list("All", "100", "50", "10")
+        )),
+        column(3, downloadButton(ns("download"), "Download"))
+      ),
+      ...,
+      DT::dataTableOutput(ns("data"))
+    )
   )
 }