+++ NEWS.md
@@ -1,6 +1,7 @@
 # hcwstudyapp 0.2.1-dev
 
 * Added scheduled dates to the participant table
+* Added variable groups to the participant table
 
 # hcwstudyapp 0.2.1